Added support for simulating device orientation changes
This functionality can be tested using the DirectorTest. The test
assumes the device starts in portrait, so you'll need to click 'rotate
device' twice the first time. It will also place the app in portrait
mode when returning to the main menu. Those quirks will, of course,
disappear when you manage your own orientation changes in your app.